extracted Position.ts to separate class
