+++ frontend/src/app/maze-builder/maze-builder.component.ts
@@ -1,22 +1,6 @@
 import {Component} from '@angular/core';
 import {MazeBuilderAutoSolveService} from "../maze-builder-auto-solve.service";
-
-export class Position {
-  x: number;
-  y: number;
-
-  constructor(x: number, y: number) {
-    this.x = x;
-    this.y = y;
-  }
-
-  public equals(obj: Object): boolean {
-    if (obj instanceof Position)
-      return obj.x == this.x && obj.y == this.y;
-
-    return false;
-  }
-}
+import {Position} from "../Position";
 
 
 @Component({
